+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The `gltf` crate adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Renamed `validate_minimally` as `validate`.
+- Removed depedency on `cgmath`.
 
 ### Fixed